+++ app/models/location.rb
@@ -5,6 +5,9 @@ class Location < ApplicationRecord
   has_many :persons, dependent: :nullify
   has_many :phone_numbers, dependent: :destroy
 
+  accepts_nested_attributes_for :phone_numbers, allow_destroy: true,
+    reject_if: :all_blank
+
   validates :name, presence: true, uniqueness: true
   validates :kind, presence: true